add warp and roll test state tests
