+++ crates/forge/tests/integration/mod.rs
@@ -1,3 +1,4 @@
+mod cheat_fork;
 pub(crate) mod common;
 mod declare;
 mod deploy;
