+++ crates/forge/tests/integration/setup_fork.rs
@@ -154,53 +154,3 @@ fn fork_cairo0_contract() {
 
     assert_passed!(result);
 }
-
-#[test]
-fn prank_cairo0_contract() {
-    let test = test_case!(formatdoc!(
-        r#"
-            use starknet::{{class_hash::Felt252TryIntoClassHash, contract_address_const,
-                SyscallResultTrait}};
-            use snforge_std::{{start_prank, stop_prank, test_address}};
-            use debug::PrintTrait;
-
-            const CAIRO0_CLASS_HASH: felt252 = 3390338629460413397996224645413818793848470654644268493965292562067946505747;
-            const LIB_CALL_SELECTOR: felt252 = 219972792400094465318120350250971259539342451068659710037080072200128459645;
-
-            #[test]
-            #[fork(url: "{}", block_id: BlockId::Number(314821))]
-            fn test() {{
-                let caller = starknet::library_call_syscall(
-                    CAIRO0_CLASS_HASH.try_into().unwrap(),
-                    LIB_CALL_SELECTOR,
-                    array![].span(),
-                ).unwrap_syscall()[0];
-
-                start_prank(test_address(), 123.try_into().unwrap());
-
-                let pranked_caller = starknet::library_call_syscall(
-                    CAIRO0_CLASS_HASH.try_into().unwrap(),
-                    LIB_CALL_SELECTOR,
-                    array![].span(),
-                ).unwrap_syscall()[0];
-
-                stop_prank(test_address());
-
-                let unpranked_caller = starknet::library_call_syscall(
-                    CAIRO0_CLASS_HASH.try_into().unwrap(),
-                    LIB_CALL_SELECTOR,
-                    array![].span(),
-                ).unwrap_syscall()[0];
-
-                assert(*pranked_caller == 123, 'start_prank does not work');
-                assert(unpranked_caller == caller, 'stop_prank does not work');
-            }}
-        "#,
-        CHEATNET_RPC_URL,
-    )
-    .as_str());
-
-    let result = run_test_case(&test);
-
-    assert_passed!(result);
-}
